requirements.txt
# Core HTTP + env loading
requests>=2.31.0
python-dotenv>=1.0.1

# Progress bar (used by assess_csv.py)
tqdm>=4.66.0
